🚚 move: TextDefault & NumberDefault & NumberCompare 분리
