--- app/src/components/elements/DashboardContentView/Text/index.ts
@@ -1,3 +1,0 @@
-export * from './NumberCompare';
-export * from './NumberDefault';
-export * from './TextDefault';
