+++ app/src/pages/HomePage/contents/Eval/AverageCommentLength.tsx
@@ -5,7 +5,7 @@ import {
   ApolloBadRequest,
   ApolloNotFound,
 } from '@components/elements/DashboardContentView/Error';
-import { NumberDefault } from '@components/elements/DashboardContentView/Text';
+import { NumberDefault } from '@components/elements/DashboardContentView/NumberDefault';
 import { DashboardContent } from '@components/templates/DashboardContent';
 
 const GET_AVERAGE_COMMENT_LENGTH = gql(/* GraphQL */ `
