+++ app/src/pages/HomePage/contents/User/MonthlyBlackholedCount.tsx
@@ -6,7 +6,7 @@ import {
   ApolloBadRequest,
   ApolloNotFound,
 } from '@components/elements/DashboardContentView/Error';
-import { NumberCompare } from '@components/elements/DashboardContentView/Text';
+import { NumberCompare } from '@components/elements/DashboardContentView/NumberCompare';
 import { DashboardContent } from '@components/templates/DashboardContent';
 import dayjs from 'dayjs';
